Resolves issue #814
Product specifications may now update dynamically based on the currently
selected options if they result in a specific matrix entry.

To allow a spec to be affected, it must be added to the $matrix_fields
array in cubecart.class.php and an HTML element with "spec_{field_name}"
as its id added to the content.product.php template.

Currently implemented are: product_code, stock_level, and product_weight;
upc, ean, jan, and isbn are pulled from the matrix, but not displayed.