Issue #902: Update catalogue pricing logic
This seems like it should go with the last commit to update pricing logic, but I can't see that it has any effect anywhere.

Also, the logic in this section differs from that in cart.class.php in that an option with an absolute price ignores any potential 'negative' modifier, whereas the logic in cart.class.php attempts to respect the 'negative' setting. This remains consistent with the original behavior, intentional or not.

All code handling pricing could probably do with some serious review to ensure a. correctness and b. consistency. Might help to make a flow chart or something detailing how it is supposed to work.