Issue #902 - fix mini-basket display using a different price calculation
Actual cart content pricing as seen in cubecart.class.php#_displayBasket sets these 2 values to the product's actual price and price * quantity, whereas the mini-basket is trying to use the 'base' product price from before any option modifiers and then adding the sum of those modifiers later.

Assuming the contents of the basket as available in the cubecart class is
the same as the one created using the Cart class, and seeing how the cart
prices, in testing, were the expected values whereas the mini-basket's were sometimes too high, it would make sense to follow the same logic that the cart uses in both cases.

This results in consistent price values from the product page to the mini-basket to the shopping cart (and, presumably, payment gateways).