Issue #902 - added price and sale price to option matrix
This allows matrix entries to be individually on sale or sold for a price that differs from what the combined options would otherwise produce. Both prices are considered absolute, taking precedence over any other absolute price that may be present in the included options.

Non-matrix options with price modifications are included normally.

Note that using absolute pricing on non-matrix options in addition to specifying one or more matrix entry prices is not recommended due to the ambiguity it introduces; as a precaution, in such cases the highest absolute price will be used as the base.