Issue #881 - 'Option_price_ignoring_tax' is actually the total price **including** tax
'Price' and 'sale_price' have had tax removed, but if tax was not included
to begin with, they will be the same as 'option_price_ignoring_tax' which
I have renamed to 'price_including_tax'.

Note also that this value already includes any modifiers from options when
assigning the 'line_price_display' and 'price_display' values, so that
code has been simplified to reflect that.