Issue #881 - establishing the groundwork
Added method to 2.cubecart.js that sends the selected options via AJAX to
the server for processing; the server then responds with the updated
product information. The previous method `price_inc_options` is left
untouched for legacy purposes.

This new AJAX call lays the foundation for updating the displayed data for
a product in real time as a customer selects options on the product page,
as well as providing a framework for integrating matrix-level data such as
pricing and product codes.

Note that there are some slight changes to the way option pricing is
handled:

1. Absolute prices: previously, if an option with an absolute price was selected, any price modifications from options appearing prior were ignored, while those from options appearing after were applied. Now, all price modifications from all options are applied on top of the absolute price.

Alternatively, we can opt to not apply any price modifications in such cases, but we should at least consider allowing modifications from non-required options to allow for truly optional options to affect price, e.g. a t-shirt with an absolutely priced option as well as an additional option to add a name tag for +$5.00.

2. Multiple absolute prices selected: previously, the price for the option
that displays last on the page was used; now, the highest of all selected
absolute prices is used instead. While it is arguable that only one option
with an absolute price should be present, this change helps keep the price
from being completely dependent on layout in the event that a store
manager decides they need multiple absolute prices.